add vllm_utils
