+++ tests/vllm_utils/setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup
+
+setup(
+    name='vllm_utils',
+    version='0.1',
+    packages=['vllm_utils'],
+)
